authority-dashboard/requirements.txt
# ---------- App ----------
streamlit==1.28.0
streamlit-folium==0.15.0
folium==0.15.0
plotly==5.17.0

# ---------- Geo ----------
geopandas==0.14.0
shapely==2.0.2

# ---------- ML ----------
xgboost==2.0.3
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.26.2
pandas==2.1.3

# ---------- LLM / Agents ----------
groq>=0.5.0
langchain>=0.1.16,<0.2.0
langchain-groq>=0.1.0

google-generativeai==0.3.2

# ---------- Utilities ----------
requests==2.31.0
python-dotenv==1.0.0
fastapi
uvicorn
pydantic